Fix `Style/OneClassPerFile` cop - gif reader
